refactor: move plugin menu data to core
