+++ packages/editor/src/package/utils/get-icon-element.tsx
@@ -19,7 +19,7 @@ import IconText from '@editor/editor-ui/assets/plugin-icons/icon-text.svg'
 import IconVideo from '@editor/editor-ui/assets/plugin-icons/icon-video.svg'
 import IconImageGallery from '@editor/editor-ui/assets/plugin-icons/image-gallery/icon-image-gallery.svg'
 
-import { PluginMenuType } from '../plugin-menu-data'
+import { PluginMenuType } from '../../core/plugin-menu-data'
 
 const iconLookup: Record<PluginMenuType, JSX.Element> = {
   [PluginMenuType.Text]: <IconText />,
